+++ dashboard/src/App.tsx
@@ -1,5 +1,6 @@
-import { Routes, Route, Navigate } from 'react-router-dom'
+import { Routes, Route } from 'react-router-dom'
 import { AppShell } from './components/AppShell'
+import { HomePage } from './pages/HomePage'
 import { RuleCatalogPage } from './pages/RuleCatalogPage'
 import { ExamplesPage } from './pages/ExamplesPage'
 
@@ -7,7 +8,7 @@ export default function App() {
   return (
     <AppShell>
       <Routes>
-        <Route path="/" element={<Navigate to="/rules" replace />} />
+        <Route path="/" element={<HomePage />} />
         <Route path="/rules" element={<RuleCatalogPage />} />
         <Route path="/examples" element={<ExamplesPage />} />
       </Routes>